chore: move some code around
